Methods distribution simplification
