+++ core/src/main/java/io/github/lemon_ant/jharmonizer/core/sorter/spoon/dependency_graph/ExplicitDeclaringTypeInitializerFieldDependencyProvider.java
@@ -1,7 +1,12 @@
 package io.github.lemon_ant.jharmonizer.core.sorter.spoon.dependency_graph;
 
+import static io.github.lemon_ant.jharmonizer.core.sorter.spoon.dependency_graph.OrderDependentFieldReferenceUtils.requireDeclaringType;
+
 import lombok.NonNull;
+import spoon.reflect.code.CtFieldAccess;
+import spoon.reflect.code.CtTypeAccess;
 import spoon.reflect.declaration.CtField;
+import spoon.reflect.declaration.CtType;
 import spoon.reflect.declaration.ModifierKind;
 
 /**
@@ -27,6 +32,23 @@ final class ExplicitDeclaringTypeInitializerFieldDependencyProvider
 
     @Override
     protected boolean hasExplicitReferenceTo(@NonNull CtField<?> referrerField, @NonNull CtField<?> referencedField) {
-        return hasExplicitDeclaringTypeReferenceTo(referrerField, referencedField);
+        CtType<?> referrerDeclaringType = requireDeclaringType(referrerField);
+
+        return hasExplicitQualifiedReferenceTo(
+                referrerField,
+                referencedField,
+                (fieldAccess, ignoredDeclaringType) ->
+                        isExplicitDeclaringTypeQualifiedAccess(fieldAccess, referrerDeclaringType));
+    }
+
+    @SuppressWarnings("PMD.CompareObjectsWithEquals")
+    private static boolean isExplicitDeclaringTypeQualifiedAccess(
+            CtFieldAccess<?> fieldAccess, CtType<?> declaringType) {
+        if (!(fieldAccess.getTarget() instanceof CtTypeAccess<?> typeAccess) || typeAccess.isImplicit()) {
+            return false;
+        }
+
+        return typeAccess.getAccessedType() != null
+                && typeAccess.getAccessedType().getTypeDeclaration() == declaringType;
     }
 }
