+++ core/src/main/java/io/github/lemon_ant/jharmonizer/core/sorter/spoon/dependency_graph/ExplicitThisInitializerFieldDependencyProvider.java
@@ -1,7 +1,10 @@
 package io.github.lemon_ant.jharmonizer.core.sorter.spoon.dependency_graph;
 
 import lombok.NonNull;
+import spoon.reflect.code.CtFieldAccess;
+import spoon.reflect.code.CtThisAccess;
 import spoon.reflect.declaration.CtField;
+import spoon.reflect.declaration.CtType;
 
 /**
  * Provides declaration dependencies created by explicit {@code this.<field>} references in field initializers.
@@ -21,6 +24,13 @@ final class ExplicitThisInitializerFieldDependencyProvider
 
     @Override
     protected boolean hasExplicitReferenceTo(@NonNull CtField<?> referrerField, @NonNull CtField<?> referencedField) {
-        return hasExplicitThisReferenceTo(referrerField, referencedField);
+        return hasExplicitQualifiedReferenceTo(
+                referrerField,
+                referencedField,
+                ExplicitThisInitializerFieldDependencyProvider::isExplicitThisQualifiedAccess);
+    }
+
+    private static boolean isExplicitThisQualifiedAccess(CtFieldAccess<?> fieldAccess, CtType<?> ignoredDeclaringType) {
+        return fieldAccess.getTarget() instanceof CtThisAccess<?> thisAccess && !thisAccess.isImplicit();
     }
 }
